requirements.txt
pyrofork
tgcrypto-pyrofork
dnspython==2.6.1
Quart==0.19.6
pymongo==4.4.1
PyYAML==6.0
requests-toolbelt==1.0.0
regex
python-dotenv
pycryptodome